# 8334843: RISC-V: Fix wraparound checking for r_array_index in lookup_secondary_supers_table_slow_path